PipelineMiRNA: Add uppercase option to parse_multi_fasta
For the special case of filtering MiRBase, we need to parse a
multi-FASTA file while uppercasing the sequence short name.
Although this is very specific, we did not want to duplicate
the functionnality of parse_multi_fasta, so we patched the method
to have an optional argument to trigger the uppercase.