PipelineMiRNA: Add static HTML files and wrapper methods
HTML header, menus and footer are always the same for all CGI scripts.
We do not want to repeat ourselves in serveral scripts.

HTML snippets are defined in static text files and transcluded from
CGI scripts using wrapping methods provided in module WebTemplate.

(This is a temporary solution pending a real templating system).