PipelineMiRNA: Add example sequence feature
As a user, we want to quickly test the pipeline using
some example sequences.

The example sequence feature solves this by providing in the interface
an easy way to add a couple of example sequences.

Add a JavaScript method to add a sequence example to the text area.
Add an interface anchor to trigger it.

In this implementation, the example sequences is stored in the
JavaScript - it might be better to externalise it to a text file.

The interface anchor used is a HTML link. It might be better to use
a different UI element such as button.