PipelineMiRNA: Revamp the logging architecture
As a developer/maintener, we want to log many events across modules.

In this commit, we store the logging file handle at the package level.
This allows us to factorise the logging setup between different scripts
(in different Perl execution, thus not possibly sharing the filehandle).
Also, we make use of the global variable setup of Log::Message to
be able to write from different modules.