PipelineMiRNA: fix bug with exonerate execution
The execution of exonerate was failing on the remote server
(but not on local testing).
(Initial bug report was a crash when displaying the results page,
which turned out to be caused by an error during YAML parsing, due
to an empty YAML file (meant to be produced by exonerate).

Investigation showed that the empty file could have no other reason
that a failure of the system command itself.
Running the command by itself returned
`/../bin/exonerate: No such file or directory`
even though the executable was indeed there.

This commit solves the issue by first cd-ing into the 'bin'
directory, and from there executing exonerate.

It remains unclear whether the original issue lies with the exonerate
binary, the Perl code (unlikely) or the system shell.