PipelineMiRNA: Start to add ODF export feature
As a user, we want to generate a full report of the pipeline
results and download it.

The implementation chosed for this feature request is to generate
a file in the OpenDocumentFormat (ODF).

We use for this the ODF::lpOD library, which must be installed
on the server.
A new module OpenDocument.pm provides methods to generate the report
given a job Identifier, and write it on disk.

Interface wise, add a anchor as a ODF image.
This triggers a JavaScrtipt method which loads a Perl script, which
in turn call the relevant OpenDocument.pm method, the refreshes
to the file location.
The user has his browser downloading the ODF file.

Future work would entail generating the actual contents of the ODF.
Possible improvements would be to avoid writing the report on disk
(it might be possible to only generate it on the fly to the browser).