PipelineMiRNA: Add GFF export feature
As a user, we want to export the pipeline results as a GFF file
(a standard file format for storing genomic features in a text file).

This commit implements a barebone GFF export feature.

A new module GFF.pm provides methods to generate the GFF
given a results structure and a job Identifier.

Interface wise, add a anchor as a download image.
This triggers a JavaScript method which loads a Perl script, which
in turn call the relevant GFF.pm method and returns an empty HTML page
as an attachment.

The generated file is GFF valid but may not be complete.
The current interface anchor is a quick hack to make things going,
we will need to provide a better interface.