PipelineMiRNA: Tweak display of alignements
As a user, we want to have a representation of the alignments
(produced by exonerate) in order to quickly visualise this information.

The exonerate output (in YAML) is parsed by the method parse_exonerate_alignment
which produces an ASCII alignment.

In this commit, we fix the output so that there are no spaces between each base,
making the result way more clearer.