PipelineMiRNA: Tweak display of hairpin
As a user, we want to have a representation of the hairpin.

The make_ASCII_viz method takes a structure and a sequence and
builds an ASCII representation of the structure.

In this commit, we fix the output so that there are no spaces between each base,
making the result way more clearer.