PipelineMiRNA: Move code to library in resultsRow
As a developer, we want to have most of the code in modules rather
than in scripts, in order to lower the complexity of the code and
enable code reuse.

In the script resultsRow, a piece of code makes a nicer output
for Vienna+Sequence couple (by cutting lines too long).

In this commit we make a subroutine out of it, and move it
to the library in WebFunctions.pm.