PipelineMiRNA: Fix bug in GFF export
Previously, we changed the signature of the method generate_GFF
to deal directly with results (as an hash), and created
generate_GFF_from_ID which takes the ID, retrieves the results based
on it and generates returns the GFF from the other method.

But we forgot to update the script cgi-bin/resultsAsGFF.pl
Fixing that.