PipelineMiRNA: Use image bare path in results structure
When parsing the job results directory, we fill a data structure with
all the information. We used to store the path to the image as a
"server path", ie the path to the image prepended with the server root.
But in some cases (like the ODF export) we need the absolute path
(ie, prepended with the path on the actual filesystem).

Change method retrieve_candidate_information to return the bare path
to the image.
Change method resultstruct2pseudoXML to make it a server path
(note the hoops we are jumping through, as the DOM changing JavaScript
*requires* that the XML-like elements must be in a specific order...)