PipelineMiRNA: Expand OpenDocument export
Expand the OpenDocument export with results contents.
We have a level 2 header for each candidate, and bare-formatted
information.

Image insertion is still shaky, as only one image get included in
the final ODF. This might be because the all have the same name
(should be investigated further though).

Possible solution would be to copy and rename images to some temporary
location before inserting them.