PipelineMiRNA: Fix Vienna parsing
Vienna parsing was broken in an earlier commit by a careless
regex editing, which captured the linebreak as well.
For the record, this surfaced in the CSV export, since it completely
broke the output format.

Fix the regex to not capture the final line break, and expand
the regex in a readable format in the process.
Possibly breaking change: we no longer capture the score together
with the dot-bracket notation. If useful, this should be carried over
in the structure.