PipelineMiRNA: Change argument passing for resultsRow page
Currently, results elements (p_value, vienna, etc.) are passed to
results page in the URL. This is quite evil, and forbids us to
retrieve anything else.

Edit the JavaScript to pass instead the JobID, sequence name and
positions (from which the candidate name can be inferred).
With this information, we can reconstruct all the candidate result
structure.