PipelineMiRNA: Guarding retrieve_candidate_information
The method which retrieves every piece of candidate information,
based on the job, the sequence name and the candidate name,
needs to be guarded against bad input.

Here, we check if the fully constructed path exists, and if so,
we throw an exception (using die).

For more clarity, we split the method in two, and tweak their signatures.
- The new one operates on the fully constructed path and contains most
  of the actual code
- The old one was reduced to only make the checking, call the other one,
  and fill in the results details which cannot be inferred from the full path.