PipelineMiRNA: Add Vienne export
As a user, I want to retrieve the sequence and structure information
of a given candidate as a file in Vienna format.

The new script, exportVienna.pl, takes as GET arguments the jobId,
candidate name and candidate position, retrieve the candidate information
and formats the sequence and structure as a Vienna file.

Add to resultsRow a link to the script.