PipelineMiRNA: Fix bug in hairpin ASCII visualisation
To display the sequence as a hairpin, we need to know where to stop
aligning (as there is nothing left to align), and to start the final loop.
The stop condition was completely baroque. We only need to check
if we considered as many parentheses as there are.

Adjust unittest to match the new code.

Note the final loop display is not great when its length is three.