PipelineMiRNA: Add functions for binaries
As a user, I want the pipeline to warn me if I am missing any dependencies,
not failing miserably with no explanation, in order to be able to fix the
problem.

Add two methods, which list the necessary 3rd party programs,
and list the missing ones. This will be useful to answer this user story.