PipelineMiRNA: abort exec if missing deps
As a user, I want the pipeline to warn me if I am missing any dependencies,
not failing miserably with no explanation, in order to be able to fix the problem.

Before launching the pipeline scripts, we check if there are any missing
3rd party binary dependencies (as listed in Programs.pm), and if so,
we redirect the user to an error page with a (hopefully) helpful
error message.