PipelineMiRNA: Revamp way results dir is set
As a developer, I need to be able to locate the results directory
consistently among needs.

Before, it was implicit that the results directory is called 'results'
and that it is situated just below the program root.
In this commit, the directory name is moved to the paths config file.
It is still assumed that the results directory is located relatively
to the program root.

A dedicated method, get_results_dir_name(), retrieves it.
It is used by the method which build the job paths (which is the main
use case).
But the direct access method will also be useful for retrieving the results
location only (ie not the current result folder below)