PipelineMiRNA: Add failure with unwritable results dir
As a user, I want to have a meaningful explanation when the pipeline
cannot operate because of file permissions, in order to be able to fix
it or ask my sysadmin to do so.

Before, this problem manifested when trying to set up the log file,
which failed because the web user could not write into the designated
results directory (and only if fatalsToBrowser were enabled, and they should not be).

Here, right at the beginning of the process (before launching the actual pipeline),
we  retrieve the results directory using the new dedicated method and check if it
is writeable. If not, we redirect to the failure page with a meaningful error message
and throw a die exception with the same error message.