PipelineMiRNA: Sort results by name/position
As a user, I want to have the pipeline results sorted by name
then position, in order to more easily browse the results.

The results are stored in an associative array key=>value where
key is the concatenation of the sequence name and the position,
which is exactly what we need to sort on.

Instead of mindlessly iterating over the hash, we first retrieve
the keys and sort them, and iterate over those.
We do this in the pseudo-XML (converted to HTML later on),
as well as the CSV, GFF and OpenDocument formats.