PipelineMiRNA: Early failure if results dir missing
As a user, I want to have a meaningful explanation when the pipeline
cannot operate because of unexisting directory, in order to be able to fix
it or ask my sysadmin to do so.

Here, we check if the results directory exists before running
the pipeline, and if not we redirect to the error page and
throw a die exception.