PipelineMiRNA: Add dummy quality computation
As a user, I want to have a quality evaluation of each candidate,
in order to quickly spot the most promising ones.

The forseen implementation is to have a star-based system based on given
criteria:
- a MFEI greater than X gives one star
- a length between 80 & 200 given another star
- the existence of a hit with MirBase another one

Final values are not yet determined.
Here we implement a skeleton of this, with a new method which computes
a score based on said candidate features.

We also export it to the pseudo XML