PipelineMiRNA: Fix bug in candidate info retrieving
In a previous commit, we changed the signature of the method
retrieve_candidate_information in module WebFunctions, to use
the object syntax (with a self first parameter). This was needed
after a reorganisation of other methods in the module.

It was thought that candidate retrieving occurred only when retrieving
all results, overlooking the three cgi-bin scripts which directly call
the method.

Thus, this change broke the row report and the cell report.
Opening the page failed with claiming that no candidate existed
for the given information.

Here, we fix the syntax of the method call for the object like one.