PipelineMiRNA: Retire outVienna thingies
The output of RNAfold (whose format is <name>\n<seq>\n<struct> (<score>))
was processed basically to output the same without the score and on one line.
It turns out this format was used nowhere in the actual pipeline, and only
during the results retrieving to get the structure and sequence, where it can
totally be retrieved from the RNAfold output. This was thus quite useless.

Here, we remove the method that does that, remove its call, and refactor
candidate retrieving to parse instead the RNAfold output.
We change a bit the signature of RNAfold parsing method to use files
rather than filehandles, per consistency.