PipelineMiRNA: Update unittest for hairpin viz
We previously had a bug regarding the hairpin ASCII visualisation,
with not well formed structures. It turned out it was because we
used the optimal structure, not the hairpin structure as a base.
This was fixed in a previous commit as part of the restructuration
of the Vienna structure retrieving (not relying anymore on the outVienna
file but reading directly from outRNAfold).

Here we update the unittest which was supposed to track this bug.