PipelineMiRNA: Add safeguard for alignment display
As a user, I want clean error messages, not frightening screens of deaths.

To display the alignment, the url of the alignment file is passed along
in GET (which in itself is quite debatable…), then the fie is parsed.
Here, we check if the URL actually points somewhere real before trying to
parse it.