PipelineMiRNA: Fix bug in 3rd party safeguard
In 6118c86 & c284abf (r8348 & r8342) we introduced a safeguard
with which the pipeline is not executed if we are missing any
3rd part binary dependencies.

We made that by having a list of the expected paths to said binaries,
and checked whether these paths existed (using unary operator -e).
But this actually validates paths pointing to folders, not files,
which led to bery subtle bugs.

In this commit, we use the unary operator -f to check if it is a plain file.
We could also check if the file is executable.