PipelineMiRNA: Add optimal structure retrieving
As a user, I want to have both the hairpin structure and the optimal
structure.

During the pipeline, we compute both the optimal and the hairpin structure.
Although the hairpin one is used as the base for the image and the ASCII viz.
But the optimal one is interesting too, so we parse it as well.