PipelineMiRNA: Add optimal structure display in report
As a user, I want to have in the report the optimal secondary structure
if it is different from the hairpin one.

RNAstemloop return both an haripin structure and an optimal structure,
which may or may not be different. In 9a811d3 (r8395), we added the
feature to retrieve this optimal structure, now we want to display it.

Here, in ResultsRow.pl we check if the optimal structure is different
from the hairpin one, and if so, we generate a second link to export
it as a text file in Vienna format.
For this to work, we add another GET parameter the Vienna export
CGI script for it to know which structure to retrieve and export.