Fix bug with exonerate output
Since we needed to parse the exonerate output, we are using the
'roll your own' feature to display it in a parseable format
(we chose YAML).
It appears the alignment produced by RYO is quite different from the
canonical one − it has '-' couples at the beginning, the end and at
random places. This caused the alignment to be wrong.

We add to the RYO the "Per-transition output for label", which states
whether a transition is "match", "gap" or "none".
We modify the parser to ignore the 'none' ones.
(We also change the capitalisation too).