PipelineMiRNA: Sort alignments by position
As a user, I want to have the alignments sorted by position,
in order to quickly recognise the same ones.

The data structure here is an hash 'begin-end' => [list of hits]
(where a hit is an hash too).
We want to sort by number on the begin position. The get_first_element_of_split
method retrieve splits the position by the dash and returns the first element,
and is used in the sorting creteria.