PipelineMiRNA: Sort alignment hits by score
As a user, I want the hits for a given position to be displayed
by descending score, in order to see the best ones first.

The hits are an array of hashes. We sort the array with the criteria
being a numerical comparison between the 'score' element value.
Display the score too.