PipelineMiRNA: Add method merge_alignments
As a user, I want the overlapping alignments to be merged, as they support
the same prediction.

We add a method merge_alignments which take the alignment hash and
merge with the following rule:
Given ordered positions, merge in [a..b] all [c..d] if a<=c and d<=b+2

The consensus alignment will be keyed using the MirDup score.
As this is not implemented yet, we use here as a stop-gap measure the simpler,
following criteria: the position which holds the more alignments is considered
the 'true' one. This will be changed later for the prediction score.

Add some unittests too.