PipelineMiRNA: Revamp export code, add Vienna export
As a user I want to export a selection of seqences in various formats.

The code to export only a selection of sequences is always the same
(iteration, check, export of the candidate if needed).
Here we factorise this behaviour into a unique method.
A switch statement selects the good candidate export method.
Add also code to export to Vienna dotbracket.