PipelineMiRNA: Add job config feature
As a developer, I want to know which run options were selected
by the user when launching the job.

Here we add a job config file (in INI format) which is written just under
the job directory. It is meant to be accessible as a semi-Singleton-like object
through the class PipelineMiRNA.

We write in it the job configuration.

Later on, we use this config file to trigger (or not) the test execution.