PipelineMiRNA: Add MiRdup validation to candidate report
As a user, I want to have the alignments validated by MiRdup.

Here we run MiRdup in validation mode on each candidate merged alignment
and retrieve its prediction 'true/false', and change the HTML accordingly.

Some reorgnisation too as we need now to have the candidate passed
to the make_alignments_HTML method.