PipelineMiRNA: Revamp result handling backend
The results are generated by wlaking the job filepath and parsing
the various results files everywhere. This was redone everytime
anything was needed, which was slow and just plain dumb.

Here, at the end of the pipeline execution, we parse every candidate,
stuff inside the parsing of the alignments and of MiRdup, and
serialise (in YAML) the resulting data structure on disk.

From that point, every time a candidate is needed, we just deserialise
the YAML file.