PipelineMiRNA: Display merged MirBase sequence in alignment
As a user, I want to align the candidates against a deduplicated MiRbase
and visualise this information in the report.

We now have a merged MirBase (meaning similar sequences are merged,
and their names concatenated).

When aligning with exonerate, we may hit a deduplicated sequence.
It has a lot of names in its description.

In the output file, the 'name' uses the first split on a space.
To retrieve the full header we concatenate with the description.
Then we split on the separator '|'. If there is more than one,
we then list them (with HTML links and everything) below the alignment.

Display is quite awful but could be customised with CSS.