PipelineMiRNA: Fix bug in hairpin mature coloration
As a user, I want the mature to be identified on the ASCII hairpin.

We have the start & end positions of the mature, and we need to mark the hairpin
accordingly. But the actual start position on the hairpin is not the start position,
as there may be gaps in the display.
This caused the coloring to be all wrong.

- Add to method compute_mature_boundaries the computation of
  the real start position, and return that as well
- Fix Hairpin coloration in Candidate module accordingly
- Add unittest for future reference.