PipelineMiRNA: Major revamp paths handling
As a developer/maintainer, I want the web aspect and the command-line
aspect of the pipeline to be decoupled, so I can test one without the other
(and because it is simply better too :-)

Heavily revamp the paths handling, especially the filesystem/web dichotomy.
Now the 'web' paths are reduced as much as possible, and (presumably) only
intervene for handling the  images path − ie when
1/ generating the pseudo-XML
2/ generating the OpenDocument

This allows us to test the full pipeline independently again.