PipelineMiRNA: Fix web_scripts HTML paths
We need to locate the web scripts relatively to the server root
(this was hardcoded as cgi-bin).

As a stopgap measure, we put that in the configuration file.
- Add accessor for it
- Add method in WebTemplate to build the path.
- Fix CGI scripts

But really this is the job of the server, and we should retrieve it
through $ENV{HTTP_REFERER}. But this fails miserably if a URL is
accessed directly, so we'll keep that in mind for later.