PipelineMiRNA: Fix functional tests run script
The functional tests were failing because of a difference in
the run_options.cfg config file. This file indeed contains the
timestamp of its creation.
Add it to the ignore list.

It was also failing (well, would have) because the candidate.yml
file contains the full path to the image − which obviously is expected
to be different − in our crazy setup with symlinks, but also in the test
environment.
Add option to diff to ignore this precise line
(this is -I 'fullpipeline').
Change the way diff command was executed, because it seems backticks
mess up with quotes and such.