PipelineMiRNA: Revamp results directory organisation.
This is major architectural change.

The results used to be organised in a directory tree, following
the names of the sequences on the first level, then with the positions
appended for the second level. The name of the sequence was given
by reading the FASTA header and splitting before the first whitespace.

This was a huge vulnerability: if two sequences had the same name or,
more subtle, shared the same prefix before the whitespace
(typically, "sequence 1" and "sequence 2"), then the results would be erased
again and again. Moreover, as names could be anything, they might have been
names problematic for the filesystem (though as all filesystem methods are
Perl-wrapped from the standard library, this should have been fine).

Moreover, from a more philosophical point of view, there was no real reason for
matching the filesystem organisation (which is an implementation detail) with
the semantics of the application (the sequence names and positions). All these
information are held in the sequence records (the ones in YAML).

In this change this is completely changed. The filesystem organisation for each
candidate files is now auto-numbered ("1" -> "1-1", "1-2", "2" -> "2-1" etc.)
This identifier is now the key to retrieve a given result − which implied heavy
changes in many parts of the CGI code and JavaScript code.