PipelineMiRNA: Fix miRDup key problem
Following the previous commit where we revamped the way candidates
are adressed (using IDs rather than a semantic construction), the
miRDup validation was not retrieved correctly anymore.

Quick fix to ahve a working key again using semantic names.
We might want to rather use the identifier in the future
(or not at all for that matter, positions should be enough).