PipelineMiRNA: Move MirDup computation from serialization to parsing
When computing the quality, we need the mirdup information. But this
was computed only at the serialization time (which made sense since
this is not parsing information per se, but parsing second-hand information).
We thus move the code computing mirdup earlier, before the quality computation.

Not sure if this is a stop-gap measure or a definitive one.
We might want to move the mirdup validation in the actual pipeline.