PipelineMiRNA: Add browser tests, BDD style
As a developer, I want the web application to be tested just like
the rest of the code, in order to be confident in my software.

The pipeline, as its core, is a Perl library with a few scripts
(well, let's say 60% of it). These features are tested, through
Perl unit tests for the library, and through Shell-powered functional
tests for the full scripted pipeline.

But the pipeline is also a web application, powered by Perl CGI scripts.
This should be tested as well, both to ensure non-regression and to
verify acceptance against specified features.

In this commit, we introduce browser tests through Behavioural-Driven-Development,
using Cucumber (BDD part), Rspec (assertions) and Watir (web application driving).
These are all Ruby-based software but that is not much a problem.

- mirna.feature contains three scenarios, written in Gherkin (Cucumber syntax)
  and documenting three features of the pipeline.
- mirna_steps.rb translates the Gherkin steps in Watir driving and Rspec assertions
- Gemfile and Gemfile.lock are Ruby dependency management files

Our three scenarios pass in Chromium and Firefox.
There is still one error when using the headless browser PhantomJS.

Note that these tests, at the moment, connect to bioinfotest:
- this does not test the repo content, but what is deployed on bioinfotest
- this only works if bioinfotest is available (ie from the lab).

Future work shall comprise more scenarios, maybe more browser support (IE?),
integration in our CI process, probably some tweaks in test launch (as the
current setup might be platform specific).

Note that several of these are dependant on the availability of a bioinfo clone
which would be more freely accessible (from, say, the CI server).

All in all, this is pretty awesome. Isn't it?