PipelineMiRNA: Simplify pre-pipeline sequence handling.
The sequences may be provided by the user either through the seqArea
or by uploading a FASTA file.
In both cases, sequences need to be pre-processed before launching the pipeline.
This was achieved through a variery of Perl regexes, Shell script, SED commands...

Though the process is basically the same for both providing methods,
there was a heavy duplication of code between the two.

In this commit we rationalise this aspect. Code is deduplicated,
and the dependency on the Shell script is removed, its feature being
integrated in the Perl workflow later. This saves Yet Another Parsing.

Future work may include more rationalisation, as it is probably not
useful to pre-process in several steps (and costly).