PipelineMiRNA: Move error detection to own method
As a user, if I submit incorrect input data, I want the pipeline to redirect
me to an error page explaining the good format, in order to be able to fix my mistake.

This is achieved through a regular expression on the sequence.
This code is both critical ad interesting, warranting its
extraction as a method stored in the Utils module.

Add unit tests as well.