PipelineMiRNA: Fix small bug in mask_CT_file
The mask_CT_file method takes as arguments two filenames, one being
the input file, the other the wanted output file, which is thus not
expected to exist.

The output file was opened in 'append' mode, for no good reason.
This was messing up with unit tests - at least locally, as files
are not purged ; the CI server would have been fine.