PipelineMiRNA: Fix warning in compute_alignment_quality
The method compute_alignment_quality basically adds up
the contents of $candidate{'alignment_existence'}
and has_mirdup_validation().
But the first operand may be unitialized.

This resulted in the following warning:
 Use of uninitialized value in addition (+)

Through an existence test, we fill a variable with either the contents
or zero, and use that variable in the sum.
(Perl does not have an equivalent to Python dict.get()? Really?)

(And yes, we are aiming for zero warnings. Moreover, it seems that
the warning may be the cause of some mess up with unit tests results
parsing in the Continuous Integration server.)