PipelineMiRNA: Move aligmnent post processing to MainPipeline
Alignments are post processed in two ways: merging, and validation with MiRDup.
This used to be done at serialization time, as it needed information from some parsing,
which occurs right before serialization.

This was problematic because this added a runtime dependency on the serialization/parsing
(for no good reason), and because running external programs should be in the realm
of the MainPipeline module.

Here we add to the MainPipeline a new method, post_process_alignments, which is called
right after running the alignments, if this is done (which is more logical too).
This method merges the alignments and retrive the mirdup results, and
serialize those in YAML.

These YAML files are deserialized at parsing time to be incorporated
in the candidate and serialized again.