PipelineMiRNA: Expand CLI interface
As a user, I want to run the pipeline in CLI mode,
get all the file results, as well as a HTML summary.

We introduce a new module, CLI.pm (which is a total junk room for now)
to hold all the methods needed to make it happen.

The methods do mainly two things:
- generate the outputs files for all candidates
- generate the HTML summary, which links to the above files.