PipelineMiRNA: Add step to Alert scenario
As a user, when I try to run the pipeline without providing
any sequence, then I want to have a warning and to stay on the
interface page.

In the no-sequence-alert scenario, we are supposed to stay on
the interface page after the alert.

Add step to Cucumber scenario
Add according Ruby step.