PipelineMiRNA: Use PageObject abstraction layer in browser tests
Our browser tests consist of scenarios (written in Cucumber)
which are translated into Ruby instructions, which pilot Watir.

Here we introduce another abstraction layer: page objects.
Page objects allow us to define the three pages of our applications
in a object oriented way, and to have accessors for the various elements.

This makes our Ruby step_definitions way cleaner and reusable,
and less prone to deprecation if the page layout is altered.

-Add page-objects gem to Gemfile
-Add three page objects : Interface, Waiting, and Results
-Change step_definitions to use the page objects

-Fix bug: The check if we stay on the page after the alert()
 was not working (as in, always working) - better remove it than mislead