PipelineMiRNA: Changed default behaviour of GFF export
Alter GFF export to export everything if there is nothing to filter on.

When using the GFF export, we first have a subroutine for the jobId,
which then calls another subroutine, passing along the array of sequences
to filter. This array was referenced twice, which made it needlessly hard
to get back the actual array (which we needed to retrieve its lenght).

Note how this was not a problem when comparing the sequences names,
because the Smart Match operator (~~) came back though time from
some Hellish dystopian future for sane developers, and manages to
make sense of whatver you throw at him
(see [perlop] for more horrific details).