PipelineMiRNA: Merge GFF export to Results module
There was a strong code duplication between the GFF export method
and the Fasta/Vienna ones. Overall, there was no strong reason to have
a GFF dedicated module in the first place.

Here we merge the GFF export logic to the export subroutine in Results.pm module.
Note how we add a header step, though FASTA and Vienna do not have one,
to be consistent and readable.

Add unit test with corresponding test file.

Update web script to call the new method.