PipelineMiRNA: Extract method running RNALfold on sequence
We want to run RNALfold on a given sequence. To do that we need to write
it on disk, but the hassle of doing so should be extracted into
a method.

Add method run_rnalfold which wraps run_rnalfold_on_file and takes
care of writing the given sequence on disk.