PipelineMiRNA: Fix bug in Results export
The hash representing a candidate might not contain one of the expected
values, which raises a "Uninitialized value in concat blahblah" warning.

To prevent this we were testing the value (with an if statement)
and setting the value with en empty string if the test failed.

But a zero value tests as false, and was thus replaced with an empty string.
This has huge impact since the JS transforming the XML appears to fail when
dealing with such a case, resulting in no table results displayed at all.

Here we only check if the value is 'defined'. We will see if it floats.