PipelineMiRNA: Add strand handling
As a user, I want to process both strands of the sequence, in order
to find more potentiel microRNA.

The UI part, as well as the config storing part, were handled
in previous commits ; and a lot of previous commits in MainPipeline
module were meant to pave the way for this one.

Here we concentrate on the following in the MainPipeline:
- Amend process_sequence signature so that it takes as arguments
  the studied strand information, and the total length of the sequence,
  so that they are passed along to process_RNAstemloop
- Same for process_RNAstemloop. Add the strand information to the final hash.
  In the case of the opposite strand, retrieve the positions relatively
  to the positive strand (using get_position_from_opposite_strand from Utils)
- Compute the revcomp, and launching process_sequence on it (which call the above)
- Merge the two resulting hashes and resorting them
- Add the strand information to all hashes in treat_candidates
- Add to create_directories the writing of the strand information on disk.

And also in Candidate module:
- Add strand information retrieval in parse_candidate_information

There is room for improvement here: writing the strand information on disk like this
sounds totally overkill and cumbersome. The handling of the various hashes could be
streamlined as well.

Note that this was NOT heavily tested at all, neither for algorithmic correctness
(like the position inversion) nor scientific correctness (like, the indices or
the [non]inversion of the sequence in the case of the opposite strand.