PipelineMiRNA: Allow to override URL in browser tests
The location of our application was hardcoded in the PageObject
InterfacePage on the bioinfotest URL. This is no good for testing purposes.

Here we introduce a Ruby helper module, url_module, which contains
one method. This method checks the contents of the environment
variable MIRKWOOD_URL and returns it it exists, or returns the hardcoded
bioinfotest URL otherwise.

This value is picked up by the InterfacePage.

This allows us to perform some more awesomeness as simply as:
bundle exec cucumber MIRKWOOD_URL=http://192.168.33.20/mirkwood/web_scripts/