PipelineMiRNA: Revamp programs path config handling
miRkwood uses several third party programs which have to be located
at runtime. We used to have a single $programs variable, set through
the config file pipeline.cfg, and were expecting all the programs
to be sitting there − and following some naming conventions.

Even though we do provide a Shell install script which kinda produces
this installation, it remains unlikely to work, unrealistic
as users may already have the programs for various reasons,
and overall not sophisticated enough.

Here is a first stab towards a better 3rd party program management.
We introduce a new config file, programs.cfg, which has the paths
to each expected binary (or folder expected to contain the binary,
in the case of the Vienna package).

Now users may alter the paths on a per program basis in the config file,
and do not have to fiddle in the Perl module.

(Note that it is still a long way before having proper 3rd party
programs management − especially at deployment time).