PipelineMiRNA: cd to exonerate directory only if provided
Exonerate is a funny animal: for some reason we kept failing when trying
to execute it using its full path − the solution was to first cd into
its directory, and then execute it.
(is that even working outside of Unix? Not sure. Better check that too.)

We want to accomodate the case where exonerate is in the PATH, in which case
there is no directory to change to. We do not expect the error to occur
in such a case (well who knows?) ; but we still must allow for the old behaviour.

So if the $exonerate_dir is set (through the config file) then we cd to it,
otherwise we do not.