PipelineMiRNA: Add Vagrant file to spin up VMs
As a developer, I want to be able to spin up a development environment,
so I can quickly get to work on a new environment.

We chose to use Vagrant, the super awesome tool to create and configure
portable development environments.

Vagrant spins up a VM based on the given ISO, and synchronises the current
directory to the VM.

This is just a matter of vagrant up!