PipelineMiRNA: Add provisioning using Ansible
Being able to spin up new VMs is nice, but this is a long way
from hvaing a fully complete development environment.

Vagrant allows to provision our VM using industry-standard tools.

We decided to use Ansible (<http://ansible.com>), a configuration
management software platform over SSH (written in Python).

The provisionning is written as a so-called 'playbook', in YAML,
which lists the tasks needed to deploy miRkwood.

As a further layer of abstraction and reusability, we defined 'roles'.

'mirkwood-software' aims to install all the 3rd-party software dependencies
of miRkwood. As some software dependencies are fairly complex to resolve
(meaning, they are not available under any sane packet management system),
the tasks here are further split in five task files, one per software, which
are 'included' in the main playbook.
(this also allows us to easily skip an include if the dependency was already solved)

(Note: RNAstemloop dependency *cannot* be solved at the moment, as there is no sane
way to retrieve its installation package. Let's leave it aside for the moment).

'apache' deals with the web server configuration (using Apache).
Aside form the classic 'tasks', there are two new things:
- handlers/, ie tasks which are executed if notified by another task.
- files/, which contains files to be copied to the node − in our case,
  the vhost configuration.

'browser-tests' prepares everything to run our browser tests
using Ruby, Cucumber & Watir

So, to summarise, now we can easily set up a development by spinning
up a new Virtual Machine, which is provisionned to be able to run miRkwood
− form the 3rd party software to the webserver configuration, including
our fancy test set up.

Seriously, how awesome is that?

To give it a try:
- Install Vagrant (APT repos are outdated, use <http://www.vagrantup.com>)
- Install Ansible (APT works − apt-get install ansible)
- vagrant up
- WebBrowser --> http://192.168.33.20/mirkwood/web_scripts/interface.pl

Yay!

And to run browser tests for ever more awesomeness:
- vagrant ssh
- cd /vagrant/tests-browser
- bundle exec cucumber MIRKWOOD_URL=http://192.168.33.20/mirkwood/web_scripts/ CHROME=true

\\o o// \o/